+++ lightnet4/yolo4/src/sortAlgorithms.h
@@ -8,9 +8,9 @@
 #ifndef SRC_SORTALGORITHMS_H
 #define SRC_SORTALGORITHMS_H
 
-void swap(float* a, float* b);
-int partition (float arr[], int low, int high);
-void quickSort(float arr[], int low, int high);
-void printArray(float arr[], int size);
+void swap(double* a, double* b);
+int partition (double arr[], int low, int high);
+void quickSort(double arr[], int low, int high);
+void printArray(double arr[], int size);
 
 #endif /* SRC_SORTALGORITHMS_H_ */
